fix property backing field annotations
This commit fixes an issue where KSP would drop annotations
on properties if the property is read from a compiled code and
the annotation has field target. I fixed it by also reading
annotations on the backing field for compiled classes.

Unfortunately, the way KSP reports a kotlin class coming from
source vs compiled code is still not the same. In compiled code,
annotations that target getter/setters are reported via the generated
synthetic in .class files vs they are reported on the property in
source files.

So added tests with data class and changed the test to
ignore the difference between getting a targeted annotation
on a property vs on its getter/setter.

We might want to change it going forward but either case it would probably
be difficult to distinguish. This is luckily is not the case for field
targeting annotations as they will only be visible in backing field which
is not visible in .class files.

cleanup