fix array values in annotations
This commit fixes 3 issues.
1) Kotlin implementation for annotation arrays were returning List instead
of Array. Fixed (#135).
2) Also, Java implementation was just returning a list of pairs for an array
inastead of returning it as an array. Fixed.
3) Java implementation was returning types as PsiType instead of KSType (kotlin
implementation returns KSType).

There is still an issue when array values are types. I fixed the
java implementation to find the KSType for PsiTypes but there is
an edge case where the kotlin processor drops invalid class names
which I've not implemented as it will require checking the type
of the array. Instead, java implementation returns KSErrorType.