Report INNER modifier for class desciptors
This PR fixes a bug where the inner modifier would be lost if the
kotlin sources is in dependencies (when wrapped in a descriptor).

Unfortunately, for descriptors, there is no easy way to distinguish
between JAVA_STATIC and INNER. That is, if a java nested class that is
annotated with STATIC is compiled, descriptor API will no longer report
any modifiers for it. Similarly, if a Java inner class (that has no
static annotation) is compiled, we'll start reporting INNER for it.
Even though this is slightly confusing, it is at least consistent from a
kotlin perspective.

Also fixed a bug in multi-module testing setup where we were NOT
compiling java sources in test modules.

Fixes: #231
Test: javaModifiers

more test cases

report inner for descriptors