Properly wrap PSIMethods that are constructors
This PR fixes a bug where we were wrapping all PsiMethods to
JavaMethodImpl without checking if they are constructor.

Both cases, we were wrapping them to find the containing class,
so I've added an extension function to ModuleClassResolver
to find containing descriptor via PsiMethod which abstracts
this resolution.

Also fixes a bug where ClassDescriptor.findEnclosedDescriptor was
not checking constructors (also convered in the updated tests).

Test: extended resolveJavaType and signatureMapper tests to trigger
the crash
Fixes: #265