Refactor all functions test to report by class
This is a no-op test refactor. allFunctions test was sorting all
declarations globally, making it hard to read what has changed when
a new declaration shows up.

This PR updates it to only sort in class level per group (functions,
declarations) so that the output is more clear to read and validate
visually.