Update tests to account for proper constructor in declarations
Updated Annotations test to specify the type of the annotated so that we
can distinguish between visiting the property vs constructor parameter
which also happens to be a property.

Also updated declaration test to specify the path of the declaration (if
it does not have a qName) so that we can distingush constructors'
owners.

Updated implicitElements expects to expect a real constructor instead of
a synthetic for the data class.