This PR integrates ktlint with the project
It is a fork of what androidX does with minor modifications.

There are 3 new tasks:

rootProject:ktlintApplyToIdea -> Applied ktlint configuration to the
current .idea project folder

<subproject>:ktlint -> Runs ktlint checks
<subproject>:ktlintFormat -> Fixes formatting for the project. Note that
this may not be able to handle all violations.

I've not modified the CI files to run this since it fails right now. One
of the core maintaners should run the format tasks to come to a clean
slate then we can add it to CI. I've not added a baseline file since
they are very brittle.