Improvements on JAVA_STATIC modifier detection (#373)
* Improvements on JAVA_STATIC modifier detection

This PR improves how we detect JAVA_STATIC modifier in declarations.

For descriptors, it wasn't showing up at all as the Descriptor API does
not have an API to return it. This change checks the containing
declaration's static scope to decide.

To be more consistent, I've also added checks for JvmStatic for both
descriptor and declaration implementation. I cached the ClassDeclaration
in Resolver (not great but does the job).

Also fixed another issue where we could add OPEN modifier based on
modality without checking if it is private or not. Added that logic into
the Descriptor -> Modifiers converter.

Fixes: #231
Test: javaModifiers