Don't check for JVM Static annotation
In the previous JVM_STATIC fixes PR, we added checks for JvmStatic
annotation which created misleading results as if the JvmStatic
annotated method is static in the companion, not the containing class.

This PR removes that case and updates the test to ensure we don't
dispatch JVM_STATIC from companions. We'll still dispatch it for
compiled java classes.

Fixes: #378
Test: javaModifiers.kt