Add support for getting owner class binary name for members
This PR adds a new Resolver API to get the binary class name for the
owner of a function / property declaration.
(resolver.getOwnerJvmClassName)

For top level functions/properties; this value is necessary to be able
to generate Java code.
An alternative solution would be to synthesize KSClassDeclaration for
these top level members but that carries over JVM limitations to KSP
which is not great in the long term. This solution instead provides an
API to unblock Java code generating KSP Processors without taking much of
technical debt for KMP.

Fixes: #396
Test: topLevelMembers.kt