Split Origin.CLASS into KOTLIN_LIB and JAVA_LIB
For KSAnnotationDescriptor, check whether the underlying descripter is
JavaAnnotationDescriptor.

For KSDeclarationDescriptorImpl, KSPropertyAccessorDescriptorImpl and
KSValueParameterDescriptorImpl, check whether they are contained by
JavaClassDescriptor.

For KSTypeReferenceDescriptorImpl, KSClassifierReferenceDescriptorImpl
and KSTypeArgumentDescriptorImpl, specify the origin in their creation
appropriately. Most of time it is the same as where they are created.
The only exception is when it's created for type arguemnts from
KotlinType.  In that case it is SYNTHETIC.