compiler-plugin: support withCompilation mode
In this mode, compilation happens right after processing, in the same
invocation of compiler.

In contrast, the default mode only performs processing. Compilation
needs to be invoked separately and explicitly.

This is incompatible with both incremental compilation and incremental
processing. None of them should be used with this mode.

Tested locally with kotlinc, kotlinc-js and kotlinc-native.