Don't mark nested classes as open
This CL fixes a bug in in KSP where we were marking nested
classes as public by mistake (the code seems to focus on
properties and methods of interfaces).

To keep a similar logic, I just excluded KSClassDeclarations
from that check.

I've also updated the test to be a bit more readable to only
list properties that return true by their name.

Fixes: #876
Test: declarationUtil.kt