Benchmark script fixes.
To build the benchmark runner and ksp:

`./benchmark/build.sh`

To build the TachiyomiExhaustive benchmark libs:

`ANDROID_SDK_ROOT=<path to root> ./benchmark/TachiyomiExhaustive/build.sh`

To run the benchmark:

`./benchmark/runner/runner.sh TachiyomiExhaustive ..`

The second argument to runner.sh is the relative path from the script
directory to where the Tachiyomi source and libs is located. This
will be different on the benchmark runners, so needs to be passed
in.