Incremental: Throw away results from clean inputs
One common misuse of incremental APIs is associating a non-root source,
instead of the ones obtained from root functions
(e.g., getSymbolsWithAnnotation), to an output. This non-root source can
be reached and reprocessed even when it is clean. Because it is clean,
it is not available via root functions. As a result, other outputs that
are solely based on it won't be re-generated and is deemed as removed.

Assuming that the processors are deterministic, we are throwing away
outputs from clean inputs, and recovering them from the backup as a
workaround for processors.