Be more compatible with K2
The compiler plugins are always loaded unconditioanlly regardless of K1
or K2 and causes FirKotlinToJvmBytecodeCompiler to throw. Because there
seems no way to load compiler plugins conditionally on language version,
we declare the KSP compiler plugin to be K2 compatible while doing
nothing if K2 is detected.

Tested locally with a new enough Kotlin compiler.