#1232: Include type hint into KSErrorType.
This CL makes `KSErrorType` a class with a string
"hint", that by convention should be a "simpleName"
of an unresolved type.

`KSErrorTypeClassDeclaration` is also no longer a singleton
and references a corresponding `KSErrorType`.

The implementation consists of extracting the available
info on the unresolved type on the best effort basis.

(cherry picked from commit 1e8486ea80b3d966a9c645c30156924c48c39750)