Use properties where possible in KspExtension
This PR converts all internal things in KspExtension to be properties
instead of lists and maps. This will ensure that reads are deferred for
as long as possible and will prevent forcing reads during configuration.

Unfortunately some properties cannot be changed without causing changes
to the ABI. New properties could be introduced but the names would have
to be different.