Support AGP's built-in Kotlin compilation
Previously, the sourceSet-specific KSP configurations (e.g., kspTest)
weren't being created when AGP's built-in Kotlin compilation was
enabled.

This change also increases the AGP version to 8.7.0 in order to test
against a version of AGP with the built-in Kotlin support.

This change also configures the AGP compatibility tests (AGP731IT and
AGP741IT) to run one test at a time to avoid race conditions seen when
the old AGP versions were trying to download SDKs simultaneously.

Bug: b/362279380
Test: GradleCompilationTest