Fix error when KSP plugin applied before AGP
Previously, when the KSP plugin was applied before AGP and AGP's
built-in Kotlin support for test fixtures or screenshot tests was
enabled, there would be an error when building.

The underlying cause for this error is
https://github.com/gradle/gradle/issues/31092, which caused the KSP
plugin to try to access the android extension before AGP's apply method
had completed.

This change fixes the problem by using the "com.android.base" plugin as
the indicator that a module is an Android module. The "com.android.base"
plugin is applied after the android extension has been added, so it
works around the Gradle bug.

Bug: https://github.com/google/ksp/issues/2174
Test: GradleCompilationTest
(cherry picked from commit 5b1cd4a998619a174058842fdb93e59dfc98dc46)