Fix validity check of annotations on properties
Annotations on properties may come from corresponding parameters if they
are declared in constructors. They may not be valid on properties and
need to be filtered out. The previous implementation incorrectly
excludes all annotations with @Target(AnnotationTarget.VALUE_PARAMETER),
while what should be excluded are the ones with @Target and without
@Target(AnnotationTarget.PROPERTY).