KSP2: fix the type of vararg for Kotlin sources
Previously, KSValueParameter.type was an array when it is from Kotlin
sources, while it is the declared type for Java and libraries.

For example, the KSValueParameter.type of strings is now consistently
String.

    fun (vararg strings: String)

The equivalentJavaWildcards test is updated due to a toString() change.

(cherry picked from commit 2a935515e739674b6530a9539006a7d3fb79f4fe)