Use excludedSources in Android KAPT Workaround
excludedSources prevents KSP tasks from taking dependencies on any
task that is added to the list in an effort to prevent circular
dependencies. However, if an output is added via Android Gradle
Plugin's APIs, KSP can still end up with a dependency on the
task that generates the output, potentially leading to a circular
dependency.

This change modifies the filter logic for adding outputs of Android
projects to respect outputs added to `excludedSources`, which can
be used to break the circular dependency at the cost of KSP ignoring
the added source.