Use excludedSources in Android KAPT Workaround
excludedSources prevents KSP tasks from taking dependencies on any
task that is added to the list in an effort to prevent circular
dependencies. In certain situations with AGP's API for adding a
task that has generated code in a project that still uses KAPT,
if the generated code task needs output from KSP, this can still
cause a circular dependency.

This change has the KAPT workaround code respect the excludedSources
list.